refactor styling architecture
